cli/pyproject.toml
[project]
name = "gpu-tool"
version = "0.1.0"
description = "GPUBridge CLI"
requires-python = ">=3.11"
dependencies = ["typer>=0.12.5", "httpx>=0.27.2", "python-dotenv>=1.0.0"]

[project.scripts]
gpu-tool = "gpu_tool.main:app"

[tool.ruff]
line-length = 100
